New Feature: Signature system - Several changes, see description and JIRA Issues below:
CLOSED - issue FENIX-157: Mesh the applet and the webpage. Includes:
Comunicating via Javascript;
Showing the applet in-page;
all of the tasks related with this 
https://fenix-ashes.ist.utl.pt/jira/browse/FENIX-157
OPEN - issue FENIX-176: Make a hidden input/ file input in a form which will be used to post the signature to the server via Javascript 
https://fenix-ashes.ist.utl.pt/jira/browse/FENIX-176
Incomplete - issue FENIX-114: Create a parallel signature module with the specifications agreed upon in discussion with Luis Cruz and revise and reimplement the signature system in this new format 
https://fenix-ashes.ist.utl.pt/jira/browse/FENIX-114
OPEN - issue FENIX-173: Make the applet use JavaScript calls hence the browser's session to retrieve and upload the signature content/signed data with a fallback to retrieving it by itself, to be implemented in Bennu v0.2 
https://fenix-ashes.ist.utl.pt/jira/browse/FENIX-173
OPEN - issue FENIX-159: Devise and test the XML marshalling and unmarshalling method so that we can conveniently create SignatureData classes that will efficiently gather the info needed to put in a signature. Also, devise this method taking into account the transform to XHTML 
https://fenix-ashes.ist.utl.pt/jira/browse/FENIX-159